+++ binding-generator/src/writer/rust_native/tuple.rs
@@ -77,8 +77,8 @@ impl RustNativeGeneratedElement for Tuple<'_, '_> {
 		]))
 	}
 
-	fn gen_rust_exports(&self) -> String {
-		extern_functions(self).iter().map(Func::gen_rust_exports).join("")
+	fn gen_rust_externs(&self) -> String {
+		extern_functions(self).iter().map(Func::gen_rust_externs).join("")
 	}
 
 	fn gen_cpp(&self) -> String {